[Refactor] Refactor DahsboardMenus and pei chart with custom hook
